Remove WPF and WinForms from core dependencies
