+++ Cyclops.Core/Cyclops.Core.csproj
@@ -1,10 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
 <Project Sdk="Microsoft.NET.Sdk">
     <PropertyGroup>
-        <TargetFramework>net6.0-windows</TargetFramework>
+        <TargetFramework>net6.0</TargetFramework>
         <GenerateAssemblyInfo>false</GenerateAssemblyInfo>
-        <UseWpf>true</UseWpf>
-        <UseWindowsForms>true</UseWindowsForms>
     </PropertyGroup>
     <ItemGroup>
         <Reference Include="System.ComponentModel.DataAnnotations" />
