+++ Cyclops.Xmpp.SharpXmpp/Cyclops.Xmpp.SharpXmpp.csproj
@@ -2,6 +2,7 @@
     <PropertyGroup>
         <TargetFramework>net6.0-windows</TargetFramework>
         <ImplicitUsings>enable</ImplicitUsings>
+        <UseWindowsForms>true</UseWindowsForms>
     </PropertyGroup>
     <ItemGroup>
         <InternalsVisibleTo Include="Cyclops.Tests" />
